backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pillow>=10.0
numpy>=1.26
opencv-python>=4.10
pydantic-settings>=2.4
# Optional:
# torch
# torchvision
# segment-anything
